Former-commit-id: d4684b6887512a20f2cf9de58ca7ad45573dc8a4

+++ Tools/ResourceEditor/Classes/Commands2/CommandBatch.cpp
@@ -93,7 +93,7 @@ int CommandBatch::Size() const
 
 Command2 * CommandBatch::GetCommand(int index) const
 {
-	if(index >= 0 && index < commandList.size())
+	if(index >= 0 && index < (int)commandList.size())
 		return commandList[index];
 
 	return NULL;
